add protected routes

+++ routes/recipe.routes.js
@@ -7,70 +7,80 @@ const uploadCloud = require('../config/cloudinary.js');
 
 //Show all recipes on recipe-overview
 router.get('/', (req, res, next) => {
-  Recipe.find().then((data) => {
+  if (req.user === undefined) {
+    res.redirect("login")
+  } else {
 
-    const filteredData = data.filter((recipe) => {
-      if (req.query.level && req.query.dishType && req.query.nutrition) {
-        return (
-          recipe.level === req.query.level &&
-          recipe.dishType === req.query.dishType &&
-          recipe.nutrition === req.query.nutrition
-        );
-      } else if (req.query.dishType && req.query.nutrition) {
-        return (
-          recipe.dishType === req.query.dishType &&
-          recipe.nutrition === req.query.nutrition
-        );
-      } else if (req.query.nutrition && req.query.level) {
-        return (
-          recipe.nutrition === req.query.nutrition &&
-          recipe.level === req.query.level
-        );
-      } else if (req.query.level && req.query.dishType) {
-        return (
-          recipe.level === req.query.level &&
-          recipe.dishType === req.query.dishType
-        );
-      } else if (req.query.level) {
-        return recipe.level === req.query.level;
-      } else if (req.query.nutrition) {
-        return recipe.nutrition === req.query.nutrition;
-      } else if (req.query.dishType) {
-        return recipe.dishType === req.query.dishType;
-      } else {
-        return recipe;
-      }
+    Recipe.find().then((data) => {
 
+      const filteredData = data.filter((recipe) => {
+        if (req.query.level && req.query.dishType && req.query.nutrition) {
+          return (
+            recipe.level === req.query.level &&
+            recipe.dishType === req.query.dishType &&
+            recipe.nutrition === req.query.nutrition
+          );
+        } else if (req.query.dishType && req.query.nutrition) {
+          return (
+            recipe.dishType === req.query.dishType &&
+            recipe.nutrition === req.query.nutrition
+          );
+        } else if (req.query.nutrition && req.query.level) {
+          return (
+            recipe.nutrition === req.query.nutrition &&
+            recipe.level === req.query.level
+          );
+        } else if (req.query.level && req.query.dishType) {
+          return (
+            recipe.level === req.query.level &&
+            recipe.dishType === req.query.dishType
+          );
+        } else if (req.query.level) {
+          return recipe.level === req.query.level;
+        } else if (req.query.nutrition) {
+          return recipe.nutrition === req.query.nutrition;
+        } else if (req.query.dishType) {
+          return recipe.dishType === req.query.dishType;
+        } else {
+          return recipe;
+        }
+
+      });
+
+      let recipeCounter = filteredData.length;
+      let uniqueNutrition = [];
+      data.forEach((recipe) => {
+        if (!uniqueNutrition.includes(recipe.nutrition)) {
+          uniqueNutrition.push(recipe.nutrition);
+        }
+      });
+
+      let uniqueDishtype = [];
+      data.forEach((recipe) => {
+        if (uniqueDishtype.includes(recipe.dishType) == false) {
+          uniqueDishtype.push(recipe.dishType);
+        }
+      });
+
+      res.render('recipes/recipe-overview', {
+        //The value you give to "recipes:" is the one that the view is going to receive
+        recipes: filteredData,
+        count: recipeCounter,
+        nutrition: uniqueNutrition,
+        dishType: uniqueDishtype,
+        user: req.user,
+      });
     });
+  }
 
-    let recipeCounter = filteredData.length;
-    let uniqueNutrition = [];
-    data.forEach((recipe) => {
-      if (!uniqueNutrition.includes(recipe.nutrition)) {
-        uniqueNutrition.push(recipe.nutrition);
-      }
-    });
-
-    let uniqueDishtype = [];
-    data.forEach((recipe) => {
-      if (uniqueDishtype.includes(recipe.dishType) == false) {
-        uniqueDishtype.push(recipe.dishType);
-      }
-    });
-
-    res.render('recipes/recipe-overview', {
-      //The value you give to "recipes:" is the one that the view is going to receive
-      recipes: filteredData,
-      count: recipeCounter,
-      nutrition: uniqueNutrition,
-      dishType: uniqueDishtype,
-      user: req.user,
-    });
-  });
 });
 
 router.get('/neu', (req, res, next) => {
-  res.render('recipes/add-recipe');
+  if (req.user === undefined) {
+    res.redirect("../login")
+  } else {
+    res.render('recipes/add-recipe')
+  }
 });
 
 router.post('/', uploadCloud.single('photo'), (req, res, next) => {
@@ -87,7 +97,6 @@ router.post('/', uploadCloud.single('photo'), (req, res, next) => {
   let directions = req.body.directions.split('\n'); // split new line
 
   let user = req.user;
-  console.log('Das ist der User', user)
 
   let recipe = new Recipe({
     title: req.body.title,
@@ -114,47 +123,55 @@ router.post('/', uploadCloud.single('photo'), (req, res, next) => {
 });
 
 router.get('/:id', (req, res, next) => {
-  Recipe.findById(req.params.id)
-    .then((data) => {
-      res.render('recipes/recipe-details', data);
-    })
-    .catch((error) => {
-      console.log('Error while getting the recipes from the DB: ', error);
-    });
+  if (req.user === undefined) {
+    res.redirect("../login")
+  } else {
+    Recipe.findById(req.params.id)
+      .then((data) => {
+        res.render('recipes/recipe-details', data);
+      })
+      .catch((error) => {
+        console.log('Error while getting the recipes from the DB: ', error);
+      });
+  }
 });
 
 // EDIT GET
 router.get('/:id/bearbeiten', (req, res, next) => {
-  Recipe.findById(req.params.id)
-    .then((mongoRecipe) => {
-      console.log('mongoRecipe', mongoRecipe);
-      let recipe = {
-        title: mongoRecipe.title,
-        image: mongoRecipe.image ? mongoRecipe.image : '',
-        ingredients: mongoRecipe.ingredients.join('\n'),
-        directions: mongoRecipe.directions.join('\n'),
-        level: mongoRecipe.level,
-        nutrition: mongoRecipe.nutrition ? mongoRecipe.nutrition : '',
-        portions: mongoRecipe.portions ? mongoRecipe.portions : '',
-        dishType: mongoRecipe.dishType,
-        duration: mongoRecipe.duration ? mongoRecipe.duration : '',
-        id: req.params.id,
-      };
-      console.log('newRecipe', recipe);
-      res.render('recipes/edit-recipe', { recipe });
-    })
-    .catch((error) => {
-      console.log('Error while editing the recipe from the DB: ', error);
-    });
+  if (req.user === undefined) {
+    res.redirect("../login")
+  } else {
+    Recipe.findById(req.params.id)
+      .then((mongoRecipe) => {
+        let recipe = {
+          title: mongoRecipe.title,
+          image: mongoRecipe.image ? mongoRecipe.image : '',
+          ingredients: mongoRecipe.ingredients.join('\n'),
+          directions: mongoRecipe.directions.join('\n'),
+          level: mongoRecipe.level,
+          nutrition: mongoRecipe.nutrition ? mongoRecipe.nutrition : '',
+          portions: mongoRecipe.portions ? mongoRecipe.portions : '',
+          dishType: mongoRecipe.dishType,
+          duration: mongoRecipe.duration ? mongoRecipe.duration : '',
+          id: req.params.id,
+        };
+        res.render('recipes/edit-recipe', { recipe });
+      })
+      .catch((error) => {
+        console.log('Error while editing the recipe from the DB: ', error);
+      });
+  }
 });
 
 // DELETE
 router.post('/:id/entfernen', (req, res) => {
-  console.log(req.params.id);
-
-  Recipe.findByIdAndDelete(req.params.id).then(() => {
-    res.redirect('/rezepte');
-  });
+  if (req.user === undefined) {
+    res.redirect("../login")
+  } else {
+    Recipe.findByIdAndDelete(req.params.id).then(() => {
+      res.redirect('/rezepte');
+    });
+  }
 });
 
 // EDIT POST
@@ -162,23 +179,27 @@ router.post(
   '/:id/bearbeiten',
   uploadCloud.single('photo'),
   (req, res, next) => {
-    Recipe.findOne({ _id: req.params.id }, function (err, recipe) {
-      recipe.title = req.body.title;
-      recipe.image = req.file ? req.file.url : recipe.image;
-      recipe.ingredients = req.body.ingredients.split('\n');
-      recipe.directions = req.body.directions.split('\n');
-      recipe.level = req.body.level;
-      recipe.nutrition = req.body.nutrition;
-      recipe.portions = req.body.portions;
-      recipe.dishType = req.body.dishType;
-      recipe.duration = req.body.duration;
-      recipe.save(function (err) {
-        if (err) {
-          console.error('ERROR!', err);
-        }
-        res.redirect('/rezepte');
+    if (req.user === undefined) {
+      res.redirect("../login")
+    } else {
+      Recipe.findOne({ _id: req.params.id }, function (err, recipe) {
+        recipe.title = req.body.title;
+        recipe.image = req.file ? req.file.url : recipe.image;
+        recipe.ingredients = req.body.ingredients.split('\n');
+        recipe.directions = req.body.directions.split('\n');
+        recipe.level = req.body.level;
+        recipe.nutrition = req.body.nutrition;
+        recipe.portions = req.body.portions;
+        recipe.dishType = req.body.dishType;
+        recipe.duration = req.body.duration;
+        recipe.save(function (err) {
+          if (err) {
+            console.error('ERROR!', err);
+          }
+          res.redirect('/rezepte');
+        });
       });
-    });
+    }
   }
 );
